+++ app/react/V2/Components/UI/Table.tsx
@@ -1,96 +1,20 @@
 /* eslint-disable react/jsx-props-no-spreading */
 /* eslint-disable react/no-multi-comp */
-import React, {
-  HTMLProps,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-  Dispatch,
-  SetStateAction,
-} from 'react';
+import React, { useEffect, useMemo, useState } from 'react';
 import {
   flexRender,
   getSortedRowModel,
   getCoreRowModel,
   useReactTable,
   SortingState,
-  ColumnDef,
-  CellContext,
-  Column,
-  TableState,
-  Row,
 } from '@tanstack/react-table';
-import { ChevronUpDownIcon, ChevronUpIcon, ChevronDownIcon } from '@heroicons/react/20/solid';
-
-type ColumnWithClassName<T> = ColumnDef<T, any> & {
-  className?: string;
-};
-
-type CellContextWithMeta<T, U> = CellContext<T, U> & {
-  column: Column<T> & { columnDef: ColumnWithClassName<T> };
-};
-
-interface TableProps<T> {
-  columns: ColumnWithClassName<T>[];
-  data: T[];
-  title?: string | React.ReactNode;
-  initialState?: Partial<TableState>;
-  enableSelection?: boolean;
-  onSelection?: Dispatch<SetStateAction<Row<T>[]>>;
-}
-
-const getIcon = (sorting: false | 'asc' | 'desc') => {
-  switch (sorting) {
-    case false:
-      return <ChevronUpDownIcon className="w-4" />;
-    case 'asc':
-      return <ChevronUpIcon className="w-4" />;
-    case 'desc':
-    default:
-      return <ChevronDownIcon className="w-4" />;
-  }
-};
-
-const IndeterminateCheckbox = ({
-  indeterminate,
-  className = '',
-  ...rest
-}: { indeterminate?: boolean } & HTMLProps<HTMLInputElement>) => {
-  const ref = useRef<HTMLInputElement>(null!);
-
-  useEffect(() => {
-    if (typeof indeterminate === 'boolean') {
-      ref.current.indeterminate = !rest.checked && indeterminate;
-    }
-  }, [ref, indeterminate, rest.checked]);
-
-  return (
-    <input type="checkbox" ref={ref} className={`rounded cursor-pointer ${className}`} {...rest} />
-  );
-};
-
-const CheckBoxHeader = ({ table }) => (
-  <IndeterminateCheckbox
-    {...{
-      checked: table.getIsAllRowsSelected(),
-      indeterminate: table.getIsSomeRowsSelected(),
-      onChange: table.getToggleAllRowsSelectedHandler(),
-    }}
-  />
-);
-
-// eslint-disable-next-line comma-spacing
-const CheckBoxCell = <T,>({ row }: { row: Row<T> }) => (
-  <IndeterminateCheckbox
-    {...{
-      checked: row.getIsSelected(),
-      disabled: !row.getCanSelect(),
-      indeterminate: row.getIsSomeSelected(),
-      onChange: row.getToggleSelectedHandler(),
-    }}
-  />
-);
+import {
+  TableProps,
+  CheckBoxHeader,
+  CheckBoxCell,
+  getIcon,
+  ColumnWithClassName,
+} from './TableTypes';
 
 // eslint-disable-next-line comma-spacing
 const Table = <T,>({
@@ -135,7 +59,7 @@ const Table = <T,>({
     getCoreRowModel: getCoreRowModel(),
     getSortedRowModel: getSortedRowModel(),
   });
-
+  
   useEffect(() => {
     const selectedRows = table.getSelectedRowModel().flatRows;
 
@@ -145,7 +69,7 @@ const Table = <T,>({
   }, [onSelection, rowSelection, table]);
 
   return (
-    <div className="overflow-x-auto relative">
+    <div className="relative overflow-x-auto">
       <table className="w-full text-sm text-left">
         {title && (
           <caption className="p-5 text-lg font-semibold text-left text-gray-900 bg-white">
@@ -195,7 +119,4 @@ const Table = <T,>({
     </div>
   );
 };
-
-export type { CellContextWithMeta, ColumnWithClassName, TableProps };
-
 export { Table };
